Split setup tests into focused partial files
