+++ Docs/reviewer/onboarding-unification-tracker.md
@@ -41,6 +41,7 @@ This file tracks onboarding/cleanup/maintenance unification across CLI, Web, and
 | Autodetect contract payload immutability hardening | Completed | IX CLI | Switched `SetupOnboardingAutoDetectResult` and `SetupOnboardingCheck` to `init`-based properties to reduce post-construction mutation risk. |
 | Web autodetect metadata fallback + stable status wire mapping | Completed | IX CLI Web + IX Tests | Web autodetect payload now falls back `contractVersion`/`contractFingerprint` when missing, uses explicit check-status mapping (`ok`/`warn`/`fail`), and tests now derive expected path counts from shared contract metadata. |
 | Web autodetect fail-fast enum wire guard | Completed | IX CLI Web + IX Tests | Unknown `SetupOnboardingCheckStatus` values now throw during web payload projection and regression tests assert fail-fast behavior. |
+| Setup test maintainability split (autodetect/post-apply groups) | Completed | IX Tests | Split large `Program.Setup.cs` into focused partial files (`Program.Setup.AutodetectContract.cs`, `Program.Setup.PostApply.cs`) to reduce monolith size and keep static-analysis maintainability checks actionable. |
 | Docs: canonical path matrix + Bot contract-check diagram | Completed | Docs | `Docs/reviewer/web-onboarding.md` now includes per-path auth/operation matrix and Mermaid flow for tool-driven parity verification. |
 | Docs: diagrams and flow updates | Completed | Docs | Added path-first flow docs and Mermaid diagrams in onboarding pages. |
 | Website FAQ/data updates | Completed | Website | Updated FAQ/features/how-it-works for path-first + auto-detect positioning. |
